NDRS-380: rename auction::{args => constants}
