+++ types/src/auction/constants.rs
